Set up Sphinx project and build
This Sphinx project and build is based on the setup we developed for the
original Notebook Aspect documentation at nb.lsst.io/
(https://github.com/lsst-dm/nb_lsst_io). Key design decisions:

- Use the same pip tools compile workfor we developed for SQuaRE web
apps to both make it easy to upgrade dependencies, but also ensure that
the pinned dependencies at any point in time are verified to work and
are consistent for all developers.
- Use tox to power multi-environment documentation builds, and also to
ensure that the pinned requirements are being used.
- Place the documentation root in docs/ to enable us to add extra Python
modules (as Sphinx extensions) to power and automate some of the
documentation processes.