Add a rspdocs python package
We need to add some Python code to this repo to manage Phalanx metadata
about the RSP environments, and possibly to add additional Sphinx
extensions in the future that are more specific that what might be
applicable for Docummenteer.

This allows us to switch to using pyproject.toml for the project
dependencies, dropping the requirements.txt workflow. Note that we
_could_ have pyproject.toml dependencies reference requirements.txt
files if we wanted.