Add a data model for Phalanx environment data
PhalanxEnv is a Pydantic class for storing metadata we'll need during
Sphinx build to customize content for the Phalanx environment.

PhalanxEnvService provides access to that dataset; right now it's
capable of loading the data from a JSON file. In the future we'll
directly inspect the Phalanx repository and cache that data.