Style the index pages
This is an exploration for how we might go about structuring the index
pages:

- use hidden toc trees for major landing pages
- use grid + grid-item-card to provide meaningful call-to-action buttons
to help users find the right section of the docs without reproducing the
toc tree from the left menu.

Note that we also use toctrees for some of the minor landing pages, like
the sub-indexes for /notebooks/. But in that case we limit depth to 2
and use titlesonly to avoid adding complexity to the landing page.