Add docs for contributing
These are based on the documentation from nb.lsst.io, and adapted as
needed.

Good improvements will be to document the data model for the "env" Jinja
variable (using the Pydantic model), and more systemtically document the
links and substitutions available from rst_epilog.rst.jinja.

Both the Jinja datamodel docs and epilog reference could be sub-pages of
env-specific-docs.